Refactor XmlMeta and XmlVar builders
